+++ tests/testthat/test_preprocessing.R
@@ -2,7 +2,7 @@
 set.seed(42)
 
 pbmc.file <- system.file('extdata', 'pbmc_raw.txt', package = 'Seurat')
-pbmc.test <- as(as.matrix(read.table(pbmc.file, sep = "\t", row.names = 1)), "dgCMatrix")
+pbmc.test <- as.sparse(x = as.matrix(read.table(pbmc.file, sep = "\t", row.names = 1)))
 
 # Tests for object creation (via CreateSeuratObject)
 # --------------------------------------------------------------------------------