+++ NEWS.md
@@ -1,4 +1,23 @@
-# Unreleased
+# Seurat 4.2.1 (2022-11-07)
+
+## Changes
+- Replaced import from `spatstat.core` with `spatstat.explore`
+- Fix bug in `FindMarkers()` when using `SCT` assay ([#6856](https://github.com/satijalab/seurat/issues/6586))
+
+# Seurat 4.2.0 (2022-09-21)
+
+## Changes
+- Fix legend color in `DoHeatmap()` ([#5783](https://github.com/satijalab/seurat/issues/5783))
+- Fix bug in `ScaleData()` when regressing out one gene ([#5970](https://github.com/satijalab/seurat/pull/5970))
+- Fix name pulling in `PlotPerturbScore()` ([#6081](https://github.com/satijalab/seurat/pull/6081))
+- Support spaceranger 2.0 ([#6208](https://github.com/satijalab/seurat/pull/6208))
+- Fix bug in `SpatialDimPlot()` when using `group.by` ([#6179](https://github.com/satijalab/seurat/issues/6179))
+- Add `add.noise` parameter in `VlnPlot()`
+([#5756](https://github.com/satijalab/seurat/issues/5756))
+- Fix uwot model backwards compatibility ([#6345](https://github.com/satijalab/seurat/issues/6345))
+- Allow `pseudocount.use` in differential expression functions to be set at the `Assay` level
+
+# Seurat 4.1.1 (2022-05-01)
 
 ## Changes
 - Fix `giveCsparse` related warnings in `Read10X_h5`
@@ -7,6 +26,8 @@
 - Fix `VlnPlot` to switch on rasterization only when required ([#5846](https://github.com/satijalab/seurat/pull/5846))
 - Fix `ncol` behavior in `SpatialPlot` ([#5774](https://github.com/satijalab/seurat/issues/5774))
 - Set `jitter` to FALSE in `FeatureScatter` ([#5876](https://github.com/satijalab/seurat/pull/5876))
+- Update `Cells` methods to new signature (`x, ...`)
+- Replace use of `default.stringsAsFactors()` with `getOption("stringsAsFactors")`
 
 # Seurat 4.1.0 (2022-01-14)
 ## Added